requirements.txt
facenet_pytorch==2.5.2
ffmpeg==1.4
matplotlib==3.4.3
numpy==1.21.2
opencv_python==4.5.5.64
pandas==1.4.2
Pillow==9.1.1
pytz==2021.1
scikit_learn==1.1.1
scipy==1.6.3
seaborn==0.11.2
spacy==3.1.1
toml==0.10.2
torch==1.10.0a0+3fd9dcf
torchvision==0.11.0a0
tqdm==4.62.1
vosk==0.3.42
umap==0.1.1
